+++ src/tools/illink/illink.sln
@@ -191,6 +191,12 @@ Global
 		{6D20F334-B7E4-4585-854B-8A0E2B29B4AA}.Release|x64.Build.0 = Release|Any CPU
 		{6D20F334-B7E4-4585-854B-8A0E2B29B4AA}.Release|x86.ActiveCfg = Release|Any CPU
 		{6D20F334-B7E4-4585-854B-8A0E2B29B4AA}.Release|x86.Build.0 = Release|Any CPU
+		{FF598E93-8E9E-4091-9F50-61A7572663AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FF598E93-8E9E-4091-9F50-61A7572663AE}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{FF598E93-8E9E-4091-9F50-61A7572663AE}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{FF598E93-8E9E-4091-9F50-61A7572663AE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FF598E93-8E9E-4091-9F50-61A7572663AE}.Release|x64.ActiveCfg = Release|Any CPU
+		{FF598E93-8E9E-4091-9F50-61A7572663AE}.Release|x86.ActiveCfg = Release|Any CPU
 		{3DDE7064-4B68-4979-8843-FDF4CE5A5140}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3DDE7064-4B68-4979-8843-FDF4CE5A5140}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3DDE7064-4B68-4979-8843-FDF4CE5A5140}.Debug|x64.ActiveCfg = Debug|Any CPU
