+++ src/tools/illink/test/Mono.Linker.Tests.Cases/DataFlow/GenericParameterWarningLocation.cs
@@ -6,10 +6,28 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq.Expressions;
 using System.Security.Policy;
 using Mono.Linker.Tests.Cases.Expectations.Assertions;
+using Mono.Linker.Tests.Cases.Expectations.Helpers;
 
 namespace Mono.Linker.Tests.Cases.DataFlow
 {
-	[IgnoreTestCase ("Ignore in NativeAOT, see https://github.com/dotnet/runtime/issues/82447", IgnoredBy = ProducedBy.NativeAot)]
+	// NativeAOT differences in behavior:
+	//
+	// Validation of generic parameters only matters if the instantiation can be used to run code with the substituted type.
+	// So for generic methods the validation has to happen basically always (since any access to the method can lead to the code
+	// of the method executing eventually).
+	// For generic types though the situation is different. Code on the type can only run if the type is instantiated (new)
+	// or if static members are accessed on it (method calls, or fields accesses both can lead to static .cctor execution).
+	// Others usages of the type cannot themselves lead to code execution in the type, and thus don't need to be validated.
+	// Currently linker and analyzer both validate every time there's a type occurrence in the code.
+	// NativeAOT on the other hand only validates the cases which can lead to code execution (this is partially because the compiler
+	// doesn't care about the type in other situations really).
+	// So for example local variables of a given type, or method parameters of that type alone will not cause code execution
+	// inside that type and thus won't be validated by NativeAOT compiler.
+	//
+	// Below this explanation/fact is referred to as "NativeAOT_StorageSpaceType"
+	//   Storage space - declaring a storage space as having a specific type doesn't in itself do anything with that type as per
+	//                   the above description.
+
 	[SkipKeptItemsValidation]
 	[ExpectedNoWarnings]
 	public class GenericParameterWarningLocation
@@ -19,9 +37,13 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 			TypeInheritance.Test ();
 			TypeImplementingInterface.Test ();
 			MethodParametersAndReturn.Test ();
+			MethodParametersAndReturnAccessedViaReflection.Test ();
 			FieldDefinition.Test ();
+			FieldDefinitionViaReflection.Test ();
 			PropertyDefinition.Test ();
 			MethodBody.Test ();
+			GenericAttributes.Test ();
+			NestedGenerics.Test ();
 		}
 
 		class TypeInheritance
@@ -50,7 +72,10 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 			[ExpectedWarning ("IL2091")]
 			class DerivedWithNoAnnotations<TUnknown>
 				: BaseWithPublicMethods<TUnknown>
-			{ }
+			{
+				[ExpectedWarning ("IL2091")]  // Compiler generates an implicit call to BaseWithPublicMethods<TUnknown>..ctor
+				public DerivedWithNoAnnotations () { }
+			}
 
 			[ExpectedWarning ("IL2091")]
 			class DerivedWithMismatchAnnotation<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields>
@@ -80,6 +105,14 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 				public static void GetDerivedMethods () => GetMethods ();
 			}
 
+			[ExpectedWarning ("IL2091")]
+			static void TestWithUnannotatedTypeArgument<TUnknown> ()
+			{
+				object a;
+				a = new DerivedWithMatchingAnnotation<TUnknown> (); // IL2091 due to the instantiation
+				a = new DerivedWithNoAnnotations<TUnknown> ();
+			}
+
 			public static void Test ()
 			{
 				Type t;
@@ -89,6 +122,14 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 				t = typeof (DerivedWithMismatchAnnotation<>);
 				t = typeof (DerivedWithOneMismatch<>);
 				t = typeof (DerivedWithTwoMatching<,>);
+
+				// Also try exact instantiations
+				object a;
+				a = new DerivedWithMatchingAnnotation<TestType> ();
+				a = new DerivedWithMatchingAnnotation<string> ();
+
+				// Also try with unannotated type parameter
+				TestWithUnannotatedTypeArgument<TestType> ();
 
 				DerivedWithOnlyStaticMethodReference<TestType>.GetDerivedMethods ();
 			}
@@ -136,13 +177,8 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 			}
 		}
 
-		// Method parameter instantiation doesn't technically need to be validated on its own
-		// Especially in AOT compiler the type of the method parameter won't even exist in the compilation
-		// unless something passes a real value to it, in which case that value creation will do all the necessary
-		// validations. So the declaration of the parameter type alone doesn't create an analysis hole
-		// it only becomes one if something tries to create an instance of such type.
-		// Similarly for statics - since we do validate accesses to static members, the static .ctor is covered through
-		// those as well.
+		//.NativeAOT: Method parameter types are not interesting until something creates an instance of them
+		// so there's no need to validate generic arguments. See comment at the top of the file for more details.
 		class MethodParametersAndReturn
 		{
 			class TypeWithPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
@@ -155,11 +191,11 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 
 			static void MethodWithSpecificType (TypeWithPublicMethods<TestType> one, IWithTwo<TestType, TestType> two) { }
 
-			[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+			[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 			static void MethodWithOneMismatch<TUnknown> (TypeWithPublicMethods<TUnknown> one) { }
 
-			[ExpectedWarning ("IL2091", nameof (IWithTwo<TestType, TestType>), ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
-			[ExpectedWarning ("IL2091", nameof (TypeWithPublicMethods<TestType>), ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+			[ExpectedWarning ("IL2091", nameof (IWithTwo<TestType, TestType>), ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+			[ExpectedWarning ("IL2091", nameof (TypeWithPublicMethods<TestType>), ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 			static void MethodWithTwoMismatches<
 				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields,
 				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
@@ -170,15 +206,21 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 
 			static TypeWithPublicMethods<TPublicMethods> MethodWithMatchingReturn<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods> () => null;
 
-			[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+			[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 			static TypeWithPublicMethods<TUnknown> MethodWithOneMismatchReturn<TUnknown> () => null;
 
-			[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
-			[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+			[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+			[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 			static IWithTwo<TPublicFields, TPublicMethods> MethodWithTwoMismatchesInReturn<
 				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields,
 				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
 				() => null;
+
+			class ConstructorWithOneMatchAndOneMismatch<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TMethods>
+			{
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				public ConstructorWithOneMatchAndOneMismatch (IWithTwo<TMethods, TMethods> two) { }
+			}
 
 			public static void Test ()
 			{
@@ -190,12 +232,64 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 				MethodWithMatchingReturn<TestType> ();
 				MethodWithOneMismatchReturn<TestType> ();
 				MethodWithTwoMismatchesInReturn<TestType, TestType> ();
+
+				_ = new ConstructorWithOneMatchAndOneMismatch<TestType> (null);
 			}
 		}
 
-		// Same as for method parameters - declaring a field with a specific type in itself is not problematic
-		// for the field to be "dangerous" it would need to have a non-null value and thus the instance has to be created somewhere
-		// and that point of creation is the place where the validation happens.
+		class MethodParametersAndReturnAccessedViaReflection
+		{
+			class TypeWithPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
+			{ }
+
+			interface IWithTwo<
+				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods,
+				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields>
+			{ }
+
+			static void MethodWithSpecificType (TypeWithPublicMethods<TestType> one, IWithTwo<TestType, TestType> two) { }
+
+			[ExpectedWarning ("IL2091")]
+			static void MethodWithOneMismatch<TUnknown> (TypeWithPublicMethods<TUnknown> one) { }
+
+			[ExpectedWarning ("IL2091", nameof (IWithTwo<TestType, TestType>))]
+			[ExpectedWarning ("IL2091", nameof (TypeWithPublicMethods<TestType>))]
+			static void MethodWithTwoMismatches<
+				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields,
+				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
+				(IWithTwo<TPublicMethods, TPublicMethods> two, TypeWithPublicMethods<TPublicFields> one)
+			{ }
+
+			static TypeWithPublicMethods<TestType> MethodWithSpecificReturnType () => null;
+
+			static TypeWithPublicMethods<TPublicMethods> MethodWithMatchingReturn<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods> () => null;
+
+			[ExpectedWarning ("IL2091")]
+			static TypeWithPublicMethods<TUnknown> MethodWithOneMismatchReturn<TUnknown> () => null;
+
+			[ExpectedWarning ("IL2091")]
+			[ExpectedWarning ("IL2091")]
+			static IWithTwo<TPublicFields, TPublicMethods> MethodWithTwoMismatchesInReturn<
+				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields,
+				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
+				() => null;
+
+			class ConstructorWithOneMatchAndOneMismatch<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TMethods>
+			{
+				[ExpectedWarning ("IL2091")]
+				public ConstructorWithOneMatchAndOneMismatch (IWithTwo<TMethods, TMethods> two) { }
+			}
+
+			public static void Test ()
+			{
+				// Access all of the methods via reflection
+				typeof (MethodParametersAndReturnAccessedViaReflection).RequiresNonPublicMethods ();
+				typeof (ConstructorWithOneMatchAndOneMismatch<>).RequiresPublicConstructors ();
+			}
+		}
+
+		//.NativeAOT: Field types are not interesting until something creates an instance of them
+		// so there's no need to validate generic arguments. See comment at the top of the file for more details.
 		class FieldDefinition
 		{
 			class TypeWithPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
@@ -228,10 +322,10 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 
 			class MultipleReferencesToTheSameType<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods, TUnknown>
 			{
-				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static TypeWithPublicMethods<TUnknown> _field1;
 				static TypeWithPublicMethods<TPublicMethods> _field2;
-				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static TypeWithPublicMethods<TUnknown> _field3;
 
 				public static void Test ()
@@ -246,8 +340,8 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods,
 				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields>
 			{
-				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
-				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static IWithTwo<TPublicFields, TPublicMethods> _field;
 
 				public static void Test ()
@@ -265,6 +359,57 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 			}
 		}
 
+		class FieldDefinitionViaReflection
+		{
+			class TypeWithPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
+			{ }
+
+			interface IWithTwo<
+				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods,
+				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields>
+			{ }
+
+			class SpecificType
+			{
+				static TypeWithPublicMethods<TestType> _field;
+			}
+
+			class OneMatchingAnnotation<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
+			{
+				static TypeWithPublicMethods<TPublicMethods> _field;
+			}
+
+			class MultipleReferencesToTheSameType<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods, TUnknown>
+			{
+				[ExpectedWarning ("IL2091")]
+				static TypeWithPublicMethods<TUnknown> _field1;
+				static TypeWithPublicMethods<TPublicMethods> _field2;
+				[ExpectedWarning ("IL2091")]
+				static TypeWithPublicMethods<TUnknown> _field3;
+			}
+
+			class TwoMismatchesInOne<
+				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods,
+				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields>
+			{
+				[ExpectedWarning ("IL2091")]
+				[ExpectedWarning ("IL2091")]
+				static IWithTwo<TPublicFields, TPublicMethods> _field;
+			}
+
+			public static void Test ()
+			{
+				typeof (SpecificType).RequiresNonPublicFields ();
+				typeof (OneMatchingAnnotation<>).RequiresNonPublicFields ();
+				typeof (MultipleReferencesToTheSameType<,>).RequiresNonPublicFields ();
+				typeof (TwoMismatchesInOne<,>).RequiresNonPublicFields ();
+			}
+		}
+
+		//.NativeAOT: Property types are not interesting until something creates an instance of them
+		// so there's no need to validate generic arguments. See comment at the top of the file for more details.
+		// In case of trimmer/aot it's even less important because properties don't exist in IL really
+		// and thus no code can manipulate them directly - only through reflection.
 		class PropertyDefinition
 		{
 			class TypeWithPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
@@ -298,12 +443,12 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 			class MultipleReferencesToTheSameType<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods, TUnknown>
 			{
 				// The warning is generated on the backing field
-				[ExpectedWarning ("IL2091", CompilerGeneratedCode = true, ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+				[ExpectedWarning ("IL2091", CompilerGeneratedCode = true, ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static TypeWithPublicMethods<TUnknown> Property1 {
-					[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+					[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 					get;
 
-					[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+					[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 					set;
 				}
 
@@ -313,12 +458,12 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 				}
 
 				// The warning is generated on the backing field
-				[ExpectedWarning ("IL2091", CompilerGeneratedCode = true, ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+				[ExpectedWarning ("IL2091", CompilerGeneratedCode = true, ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static TypeWithPublicMethods<TUnknown> Property3 {
-					[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+					[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 					get;
 
-					[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+					[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 					set;
 				}
 
@@ -335,14 +480,14 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields>
 			{
 				// The warnings are generated on the backing field
-				[ExpectedWarning ("IL2091", CompilerGeneratedCode = true, ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
-				[ExpectedWarning ("IL2091", CompilerGeneratedCode = true, ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+				[ExpectedWarning ("IL2091", CompilerGeneratedCode = true, ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", CompilerGeneratedCode = true, ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static IWithTwo<TPublicFields, TPublicMethods> Property {
 					// Getter is trimmed and doesn't produce any warning
 					get;
 
-					[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
-					[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)]
+					[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+					[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 					set;
 				}
 
@@ -366,8 +511,10 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 			class TypeWithPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods> : Exception
 			{
 				public static void Method () { }
+				public void InstanceMethod () { }
 
 				public static string Field;
+				public string InstanceField;
 
 				public static string Property { get; set; }
 			}
@@ -377,6 +524,7 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields>
 			{
 				public static void Method () { }
+				public void InstanceMethod ();
 
 				public static string Field;
 
@@ -390,10 +538,17 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 
 			static void MethodWithPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods> () { }
 
+			void InstanceMethodWithPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods> () { }
+
 			static void MethodWithTwo<
 				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods,
 				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields> ()
 			{ }
+
+			static MethodBody GetInstance () => null;
+
+			[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // return type // NativeAOT_StorageSpaceType
+			static TypeWithPublicMethods<T> GetInstanceForTypeWithPublicMethods<T> () => null;
 
 			class TypeOf
 			{
@@ -480,12 +635,19 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 					MethodWithTwo<TPublicFields, TPublicMethods> ();
 				}
 
+				[ExpectedWarning ("IL2091")]
+				static void InstanceMethodMismatch<TUnknown> ()
+				{
+					GetInstance ().InstanceMethodWithPublicMethods<TUnknown> ();
+				}
+
 				public static void Test ()
 				{
 					SpecificType ();
 					OneMatchingAnnotation<TestType> ();
 					MultipleReferencesToTheSameMethod<TestType, TestType> ();
 					TwoMismatchesInOneStatement<TestType, TestType> ();
+					InstanceMethodMismatch<TestType> ();
 				}
 			}
 
@@ -523,12 +685,21 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 					IWithTwo<TPublicFields, TPublicMethods>.Method ();
 				}
 
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // local variable // NativeAOT_StorageSpaceType
+				static void InstanceMethodMismatch<TUnknown> ()
+				{
+					TypeWithPublicMethods<TUnknown> instance = GetInstanceForTypeWithPublicMethods<TUnknown> ();
+					instance.InstanceMethod ();
+				}
+
 				public static void Test ()
 				{
 					SpecificType ();
 					OneMatchingAnnotation<TestType> ();
 					MultipleReferencesToTheSameMethod<TestType, TestType> ();
 					TwoMismatchesInOneStatement<TestType, TestType> ();
+					InstanceMethodMismatch<TestType> ();
 				}
 			}
 
@@ -566,15 +737,27 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 					_ = IWithTwo<TPublicFields, TPublicMethods>.Field;
 				}
 
+				// The local variable
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // access to the field // NativeAOT_StorageSpaceType
+				static void InstanceFieldMismatch<TUnknown> ()
+				{
+					TypeWithPublicMethods<TUnknown> instance = GetInstanceForTypeWithPublicMethods<TUnknown> ();
+					_ = instance.InstanceField;
+				}
+
 				public static void Test ()
 				{
 					SpecificType ();
 					OneMatchingAnnotation<TestType> ();
 					MultipleReferencesToTheSameField<TestType, TestType> ();
 					TwoMismatchesInOneStatement<TestType, TestType> ();
+					InstanceFieldMismatch<TestType> ();
 				}
 			}
 
+			//.NativeAOT: Local variable types are not interesting until something creates an instance of them
+			// so there's no need to validate generic arguments. See comment at the top of the file for more details.
 			class LocalVariable
 			{
 				static void SpecificType ()
@@ -588,8 +771,8 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 					TypeWithPublicMethods<TPublicMethods> t = null;
 				}
 
-				[ExpectedWarning ("IL2091")]
-				[ExpectedWarning ("IL2091")]
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static void MultipleReferencesToTheSameType<
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods,
 					TUnknown> ()
@@ -599,8 +782,8 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 					TypeWithPublicMethods<TUnknown> t3 = null; // Warn
 				}
 
-				[ExpectedWarning ("IL2091")]
-				[ExpectedWarning ("IL2091")]
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static void TwoMismatchesInOneStatement<
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields,
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
@@ -948,6 +1131,8 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 				}
 			}
 
+			//.NativeAOT: Checking an instance for its type is not interesting until something creates an instance of that type
+			// so there's no need to validate generic arguments. See comment at the top of the file for more details.
 			class IsInstance
 			{
 				static object _value = null;
@@ -962,8 +1147,8 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 					bool a = _value is TypeWithPublicMethods<TPublicMethods>;
 				}
 
-				[ExpectedWarning ("IL2091")]
-				[ExpectedWarning ("IL2091")]
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static void MultipleReferencesToTheSameMethod<
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods,
 					TUnknown> ()
@@ -973,8 +1158,8 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 					bool a3 = _value is TypeWithPublicMethods<TUnknown>; // Warn
 				}
 
-				[ExpectedWarning ("IL2091")]
-				[ExpectedWarning ("IL2091")]
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static void TwoMismatchesInOneStatement<
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields,
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
@@ -992,6 +1177,7 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 				}
 			}
 
+			// This is basically the same operation as IsInstance
 			class AsType
 			{
 				static object _value = null;
@@ -1006,8 +1192,8 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 					object a = _value as TypeWithPublicMethods<TPublicMethods>;
 				}
 
-				[ExpectedWarning ("IL2091")]
-				[ExpectedWarning ("IL2091")]
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static void MultipleReferencesToTheSameMethod<
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods,
 					TUnknown> ()
@@ -1017,8 +1203,8 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 					object a3 = _value as TypeWithPublicMethods<TUnknown>; // Warn
 				}
 
-				[ExpectedWarning ("IL2091")]
-				[ExpectedWarning ("IL2091")]
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static void TwoMismatchesInOneStatement<
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields,
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
@@ -1036,6 +1222,9 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 				}
 			}
 
+			//.NativeAOT: Exception types are effectively very similar to local variable or method parameters.
+			// and are not interesting until something creates an instance of them
+			// so there's no need to validate generic arguments. See comment at the top of the file for more details.
 			class ExceptionCatch
 			{
 				static void SpecificType ()
@@ -1054,8 +1243,8 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 					}
 				}
 
-				[ExpectedWarning ("IL2091")]
-				[ExpectedWarning ("IL2091")]
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static void MultipleReferencesToTheSameType<
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods,
 					TUnknown> ()
@@ -1076,8 +1265,8 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 					}
 				}
 
-				[ExpectedWarning ("IL2091")]
-				[ExpectedWarning ("IL2091")]
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static void TwoMismatchesInOneStatement<
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields,
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
@@ -1098,6 +1287,7 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 				}
 			}
 
+			// This is basically the same as IsInstance and thus not dangerous
 			class ExceptionFilter
 			{
 				static void SpecificType ()
@@ -1116,8 +1306,8 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 					}
 				}
 
-				[ExpectedWarning ("IL2091")]
-				[ExpectedWarning ("IL2091")]
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static void MultipleReferencesToTheSameType<
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods,
 					TUnknown> ()
@@ -1138,8 +1328,8 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 					}
 				}
 
-				[ExpectedWarning ("IL2091")]
-				[ExpectedWarning ("IL2091")]
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
+				[ExpectedWarning ("IL2091", ProducedBy = ProducedBy.Trimmer | ProducedBy.Analyzer)] // NativeAOT_StorageSpaceType
 				static void TwoMismatchesInOneStatement<
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields,
 					[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
@@ -1178,6 +1368,197 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 				AsType.Test ();
 				ExceptionCatch.Test ();
 				ExceptionFilter.Test ();
+			}
+		}
+
+		// There are no warnings due to data flow itself
+		// since the generic attributes must be fully instantiated always.
+		class GenericAttributes
+		{
+			class TypeWithPublicMethodsAttribute<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods>
+				: Attribute
+			{ }
+
+			class TypeWithTwoAttribute<
+				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TPublicMethods,
+				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TPublicFields>
+				: Attribute
+			{ }
+
+			[TypeWithPublicMethods<TestType>]
+			static void OneSpecificType () { }
+
+			[TypeWithTwo<TestType, TestType>]
+			static void TwoSpecificTypes () { }
+
+			public static void Test ()
+			{
+				OneSpecificType ();
+				TwoSpecificTypes ();
+			}
+		}
+
+		class NestedGenerics
+		{
+			class RequiresMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] T>
+			{
+			}
+
+			class RequiresNothing<T>
+			{
+			}
+
+			class RequiresFields<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T>
+			{
+			}
+
+			static void GenericMethodNoRequirements<T> () { }
+
+			static void GenericMethodNoRequirementsAccessToT<T> ()
+			{
+				_ = typeof (T).Name;
+			}
+
+			class GenericTypeWithMethodAndField<T>
+			{
+				public void GenericInstanceMethod<T> () { }
+				public static void GenericMethod<T> () { }
+
+				public static int Field;
+			}
+
+			class TypeWithRUCMethod
+			{
+				[RequiresUnreferencedCode("--RUCMethod--")]
+				public static void RUCMethod () { }
+			}
+
+			[ExpectedWarning ("IL2091", "TUnknown", "RequiresFields", nameof (DynamicallyAccessedMemberTypes.PublicFields))]
+			[ExpectedWarning ("IL2026", "--RUCMethod--")]
+			static void GenericMethodNesting<TUnknown> ()
+			{
+				GenericMethodNoRequirements<RequiresMethods<RequiresNothing<TUnknown>>> (); // No warning
+				GenericMethodNoRequirements<RequiresMethods<RequiresNothing<RequiresFields<TUnknown>>>> (); // IL2091
+				GenericMethodNoRequirements<RequiresMethods<RequiresNothing<RequiresMethods<TypeWithRUCMethod>>>> (); // IL2026
+			}
+
+			[ExpectedWarning ("IL2091", "TUnknown", "RequiresFields", nameof (DynamicallyAccessedMemberTypes.PublicFields))]
+			[ExpectedWarning ("IL2026", "--RUCMethod--")]
+			static void GenericMethodNestingAccessToT<TUnknown> ()
+			{
+				GenericMethodNoRequirementsAccessToT<RequiresMethods<RequiresNothing<TUnknown>>> (); // No warning
+				GenericMethodNoRequirementsAccessToT<RequiresMethods<RequiresNothing<RequiresFields<TUnknown>>>> (); // IL2091
+				GenericMethodNoRequirementsAccessToT<RequiresMethods<RequiresNothing<RequiresMethods<TypeWithRUCMethod>>>> (); // IL2026
+			}
+
+			[ExpectedWarning ("IL2091", "TUnknown", "RequiresFields", nameof (DynamicallyAccessedMemberTypes.PublicFields))]
+			[ExpectedWarning ("IL2026", "--RUCMethod--")]
+			static void GenericInstanceMethodNesting<TUnknown> ()
+			{
+				GenericTypeWithMethodAndField<TestType> instance = new ();
+				instance.GenericInstanceMethod<RequiresMethods<RequiresNothing<TUnknown>>> (); // No warning
+				instance.GenericInstanceMethod<RequiresMethods<RequiresNothing<RequiresFields<TUnknown>>>> (); // IL2091
+				instance.GenericInstanceMethod<RequiresMethods<RequiresNothing<RequiresMethods<TypeWithRUCMethod>>>> (); // IL2026
+			}
+
+			[ExpectedWarning ("IL2091", "TUnknown", "RequiresFields")]
+			[ExpectedWarning ("IL2091", "TUnknown", "RequiresMethods")]
+			[ExpectedWarning ("IL2026", "--RUCMethod--")]
+			[ExpectedWarning ("IL2026", "--RUCMethod--")]
+			static void MethodOnGenericTypeNesting<TUnknown> ()
+			{
+				GenericTypeWithMethodAndField<RequiresMethods<RequiresNothing<TUnknown>>>.GenericMethod<string> (); // No warning
+				GenericTypeWithMethodAndField<RequiresMethods<RequiresNothing<RequiresFields<TUnknown>>>>          // IL2091
+					.GenericMethod<RequiresNothing<RequiresFields<RequiresFields<RequiresMethods<TUnknown>>>>> (); // IL2091
+				GenericTypeWithMethodAndField<RequiresMethods<RequiresNothing<RequiresMethods<TypeWithRUCMethod>>>> // IL2026
+					.GenericMethod<RequiresNothing<RequiresFields<RequiresMethods<TypeWithRUCMethod>>>> (); // IL2026
+			}
+
+			[ExpectedWarning ("IL2091", "TUnknown", "RequiresFields", nameof (DynamicallyAccessedMemberTypes.PublicFields))]
+			[ExpectedWarning ("IL2026", "--RUCMethod--")]
+			static void FieldOnGenericTypeNesting<TUnknown> ()
+			{
+				GenericTypeWithMethodAndField<RequiresMethods<RequiresNothing<TUnknown>>>.Field = 0; // No warning
+				_ = GenericTypeWithMethodAndField<RequiresMethods<RequiresNothing<RequiresFields<TUnknown>>>>.Field; // IL2091
+				_ = GenericTypeWithMethodAndField<RequiresMethods<RequiresNothing<RequiresMethods<TypeWithRUCMethod>>>>.Field; // IL2026
+			}
+
+			class BaseTypeGenericNesting
+			{
+				class Base<T>
+				{
+					static Base () { _ = typeof (T).Name; }
+				}
+
+				class DerivedWithNothing<TUnknown>
+					: Base<RequiresMethods<RequiresNothing<TUnknown>>>
+				{ }
+
+				[ExpectedWarning ("IL2091", "TUnknown", "RequiresFields", nameof (DynamicallyAccessedMemberTypes.PublicFields))]
+				class DerivedWithFields<TUnknown>
+					: Base<RequiresMethods<RequiresNothing<RequiresFields<TUnknown>>>>
+				{
+					static DerivedWithFields()
+					{
+					}
+				}
+
+				[ExpectedWarning ("IL2026", "--RUCMethod--")]
+				class DerivedWithRUC
+					: Base<RequiresMethods<RequiresNothing<RequiresMethods<TypeWithRUCMethod>>>>
+				{ }
+
+				public static void Test()
+				{
+					Type a;
+					a = typeof(DerivedWithNothing<TestType>);
+					a = typeof(DerivedWithFields<TestType>);
+					a = typeof(DerivedWithRUC);
+				}
+			}
+
+			class InterfaceGenericNesting
+			{
+				interface IBase<T>
+				{
+				}
+
+				class DerivedWithNothing<TUnknown>
+					: IBase<RequiresMethods<RequiresNothing<TUnknown>>>
+				{ }
+
+				[ExpectedWarning ("IL2091", "TUnknown", "RequiresFields", nameof (DynamicallyAccessedMemberTypes.PublicFields), ProducedBy = ProducedBy.NativeAot)]
+				class DerivedWithFields<TUnknown>
+					: IBase<RequiresMethods<RequiresNothing<RequiresFields<TUnknown>>>>
+				{
+					static DerivedWithFields ()
+					{
+					}
+				}
+
+				[ExpectedWarning ("IL2026", "--RUCMethod--", ProducedBy = ProducedBy.NativeAot)]
+				class DerivedWithRUC
+					: IBase<RequiresMethods<RequiresNothing<RequiresMethods<TypeWithRUCMethod>>>>
+				{ }
+
+				public static void Test ()
+				{
+					Type a;
+					a = typeof (DerivedWithNothing<TestType>);
+					a = typeof (DerivedWithFields<TestType>);
+					a = typeof (DerivedWithRUC);
+				}
+			}
+
+			public static void Test ()
+			{
+				GenericMethodNesting<TestType> ();
+				GenericMethodNestingAccessToT<TestType> ();
+				GenericInstanceMethodNesting<TestType> ();
+				MethodOnGenericTypeNesting<TestType> ();
+				FieldOnGenericTypeNesting<TestType> ();
+				BaseTypeGenericNesting.Test ();
+				InterfaceGenericNesting.Test ();
 			}
 		}
